backend/requirements.txt
numpy>=1.20.0
librosa>=0.9.0
sounddevice>=0.4.4
soundfile>=0.10.3
noisereduce>=2.0.0
openai-whisper>=1.1.10
torch>=1.12.0
transformers>=4.16.0
pydantic>=1.8.0,<2.0.0
spacy==3.5.0
scipy>=1.8.0
cryptography>=37.0.0
python-dotenv>=0.20.0